WIP - converted to pages, implemeneted menu flyout - need to implent new page button.
