+++ CookingBook/AppShell.xaml.cs
@@ -11,7 +11,7 @@
             // needs to be registered with the navigation system.
             // The AllNotesPage and AboutPage pages are automatically registered
             // with the navigation system by being declared in the TabBar.
-            Routing.RegisterRoute(nameof(Views.NotePage), typeof(Views.NotePage));
+             Routing.RegisterRoute(nameof(Pages.NotePage), typeof(Pages.NotePage));
         }
     }
 }
