+++ CookingBook/Models/Note.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace CookingBook.Models
 {
-    internal class Note
+    public class Note
     {
         public string Filename { get; set; }
         public string Text { get; set; }
