+++ CookingBook/Pages/NotePage.xaml
@@ -1,12 +1,18 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:viewModels="clr-namespace:CookingBook.ViewModels"
-             x:Class="CookingBook.Views.NotePage"
-             Title="Note">
-    <ContentPage.BindingContext>
-        <viewModels:NoteViewModel />
-    </ContentPage.BindingContext>
+             xmlns:utes="clr-namespace:CookingBook.Utilities"
+             x:Class="CookingBook.Pages.NotePage">
+    <ContentPage.Resources>
+        <ResourceDictionary>
+            <utes:NoteTitleConverter x:Key="NoteTitleConverter" />
+        </ResourceDictionary>
+    </ContentPage.Resources>
+
+    <ContentPage.Title>
+        <Binding Path="Text" Converter="{StaticResource NoteTitleConverter}" />
+    </ContentPage.Title>
+    
     <VerticalStackLayout Spacing="10" Margin="5">
         <Editor x:Name="TextEditor"
                 Placeholder="Enter your note"
