--- CookingBook/Views/AboutPage.xaml.cs
@@ -1,9 +1,0 @@
-namespace CookingBook.Views;
-
-public partial class AboutPage : ContentPage
-{
-	public AboutPage()
-	{
-		InitializeComponent();
-	}
-}
