--- CookingBook/Views/AllNotesPage.xaml.cs
@@ -1,14 +1,0 @@
-namespace CookingBook.Views;
-
-public partial class AllNotesPage : ContentPage
-{
-    public AllNotesPage()
-    {
-        InitializeComponent();
-    }
-
-    private void ContentPage_NavigatedTo(object sender, NavigatedToEventArgs e)
-    {
-        notesCollection.SelectedItem = null;
-    }
-}
