--- CookingBook/Views/NotePage.xaml.cs
@@ -1,9 +1,0 @@
-namespace CookingBook.Views;
-
-public partial class NotePage : ContentPage
-{
-    public NotePage()
-	{
-		InitializeComponent();
-    }
-}
